Move the watching module from reactor to a K8s-API-specific package
