+++ tests/reactor/conftest.py
@@ -3,9 +3,9 @@ import asyncio
 import pytest
 from asynctest import CoroutineMock
 
+from kopf.k8s.watching import streaming_watch
 from kopf.reactor.queueing import watcher
 from kopf.reactor.queueing import worker as original_worker
-from kopf.reactor.watching import streaming_watch
 
 
 @pytest.fixture()
@@ -37,7 +37,7 @@ def worker_mock(mocker):
 @pytest.fixture()
 def watcher_limited(mocker):
     """ Make event streaming finite, watcher exits after depletion. """
-    mocker.patch('kopf.reactor.watching.infinite_watch', new=streaming_watch)
+    mocker.patch('kopf.k8s.watching.infinite_watch', new=streaming_watch)
 
 
 @pytest.fixture()
